fix install.sh to the same state with master branch

+++ install.sh
@@ -26,10 +26,10 @@ Installs sd_maskrcnn repo and other dependencies, as well as optional models and
 """
 
 # Install maskrcnn submodule
-# git submodule update --init
-# cd maskrcnn 
-# python3 setup.py install
-# cd ..
+git submodule update --init
+cd maskrcnn 
+python3 setup.py install
+cd ..
 
 # Install main module, with generation if requested
 if [ "$#" == "1" ] && [ "$1" == "generation" ]
@@ -41,21 +41,21 @@ else
 fi
 
 # Download pretrained model if requested
-# shopt -s nocasematch
-# unset response
-# while [[ ! $response =~ (y|yes|n|no) ]]; do
-#     echo -n "Download Pre-trained Model to models/ (150 MB)? (y/n) > "
-#     read response
-# done
+shopt -s nocasematch
+unset response
+while [[ ! $response =~ (y|yes|n|no) ]]; do
+    echo -n "Download Pre-trained Model to models/ (150 MB)? (y/n) > "
+    read response
+done
 
-# case "$response" in
-#     [yY][eE][sS]|[yY]) 
-mkdir -p models
-wget -v -O models/sd_maskrcnn.h5 https://berkeley.box.com/shared/static/obj0b2o589gc1odr2jwkx4qjbep11t0o.h5
-#         ;;
-#     *)
-#         ;;
-# esac
+case "$response" in
+    [yY][eE][sS]|[yY]) 
+        mkdir -p models
+        wget -v -O models/sd_maskrcnn.h5 https://berkeley.box.com/shared/static/obj0b2o589gc1odr2jwkx4qjbep11t0o.h5
+        ;;
+    *)
+        ;;
+esac
 
 if [ "$#" == "1" ] && [ "$1" == "generation" ]
 then
